+++ code/+mpqc/+settings/returnMPQC_SettingsFileName.m
@@ -1,12 +1,12 @@
- function MPSF_SettingsFileName = returnMPSF_SettingFileName
+ function MPQC_SettingsFileName = returnMPQC_SettingFileName
     % Return the mpqc settings file name as a string
     %
     % Purpose
     % Multiple files need access to this file name so we define it here once.
-    % Returns "MPSF_SystemSettings.yml"
+    % Returns "MPQC_SystemSettings.yml"
     %
     % Rob Campbell, SWC AMF, initial commit 2022
 
-    MPSF_SettingsFileName = 'MPSF_SystemSettings.yml';
+    MPQC_SettingsFileName = 'MPQC_SystemSettings.yml';
 
-end % returnMPSF_SettingFileName
+end % returnMPQC_SettingFileName